refactor(#13): parser dir
